requirements.txt
av==12.0.0
einops==0.7.0
numpy==1.26.4
Pillow==10.3.0
PyYAML==6.0.1
torch==2.2.2
torchvision==0.17.2
